Inline project git source fields
To prepare for having a different way to specify the project source,
I've taken all the fields for specifying a git repo as the project
source, put them in a struct, and embedded that struct into the spec.

This has the effect of bundling the fields into one value which can be
present or not, without affecting how YAMLs are unmarshaled. That will
help transition to having sourceRef (pointing to a Flux source) be an
alternative to giving all the git repo fields directly.

An exception is the field `RepoDir`, which I've left where it is since
it will apply to other kinds of sources.

In the expectation that some procedures will have to explicitly work
with the git repo fields, I didn't embed the inlined struct. Similarly,
helper funcs that refer to git repo fields now take an *InlineGitRepo
argument.

Signed-off-by: Michael Bridgen <mbridgen@pulumi.com>